Create Android PR when a release is created
Previously, the workflow ran whenever a tag is created in the
repository. We might create tags for all sorts of reasons, not
necessarily because we want to create a release and therefore raise a PR
in the Android repo.

This change runs the workflow to create a PR in the Android repo only
when we've created a Release in `bridget-android`. It does not run if a
prerelease was created.